Move some more tests out of tests_general
