+++ evap/contributor/tests/test_views.py
@@ -1,7 +1,7 @@
 from webtest.app import AppError
 
-from evap.evaluation.models import Course
-from evap.evaluation.tests.test_utils import ViewTest, course_with_responsible_and_editor
+from evap.evaluation.models import Course, Questionnaire
+from evap.evaluation.tests.test_utils import ViewTest, course_with_responsible_and_editor, lastform
 
 TESTING_COURSE_ID = 2
 
@@ -81,3 +81,25 @@ class TestContributorCourseEditView(ViewTest):
         course.save()
 
         self.get_assert_403(self.url, 'responsible')
+
+    def test_contributor_course_edit(self):
+        """
+            Tests whether the "save" button in the contributor's course edit view does not
+            change the course's state, and that the "approve" button does that.
+        """
+        course = Course.objects.get(pk=TESTING_COURSE_ID)
+
+        page = self.get_assert_200("/contributor/course/{}/edit".format(TESTING_COURSE_ID), user="responsible")
+        form = lastform(page)
+        form["vote_start_date"] = "02/1/2098"
+        form["vote_end_date"] = "02/1/2099"
+
+        response = form.submit(name="operation", value="save")
+        self.assertEqual(Course.objects.get(pk=TESTING_COURSE_ID).state, "prepared")
+
+        form.submit(name="operation", value="approve")
+        self.assertEqual(Course.objects.get(pk=TESTING_COURSE_ID).state, "editorApproved")
+
+        # test what happens if the operation is not specified correctly
+        response = form.submit(expect_errors=True)
+        self.assertEqual(response.status_code, 403)
